update: change file structure
