+++ package.json
@@ -6,12 +6,12 @@
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.1.1",
     "@testing-library/user-event": "^13.5.0",
-    "axios": "^0.27.2",
-    "dayjs": "^1.11.2",
-    "jwt-decode": "^3.1.2",
-    "jwt-encode": "^1.0.1",
-    "miragejs": "^0.1.43",
-    "mockman-js": "^1.1.5",
+    "axios": "0.27.2",
+    "dayjs": "1.11.2",
+    "jwt-decode": "3.1.2",
+    "jwt-encode": "1.0.1",
+    "miragejs": "0.1.43",
+    "mockman-js": "1.1.5",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
     "react-router-dom": "^6.3.0",
