+++ src/components/index.js
@@ -1,18 +1,21 @@
 
 
 import NavBar from "./navbar/Navbar";
-import HomePage from "./homepage/HomePage";
+// import HomePage from "./homepage/HomePage";
 import Sidebar from "./sidebar/Sidebar";
-import LoginPage from "./authpage/LoginPage";
-import SignUpPage from "./authpage/SignUpPage";
-import ExplorePage from "./explorepage/ExplorePage";
-import VideoPage from "./videopage/VideoPage";
+// import LoginPage from "./authpage/LoginPage";
+// import SignUpPage from "./authpage/SignUpPage";
+// import ExplorePage from "../pages/explorepage/ExplorePage";
+// import VideoPage from "./videopage/VideoPage";
 import RestrictedRoute from "./authRoutes/RestrictedRoute";
 import PrivateRoute from "./authRoutes/PrivateRoute";
-import LikedVideosPage from "./likedvideospage/LikedVideosPage";
-import WatchLaterPage from "./watchlaterpage/WatchLaterPage";
-import HistoryPage from "./historypage/HistoryPage";
-import AddPlaylistModal from "./playlistpage/AddPlaylistModal";
-import SelectPlaylistModal from "./playlistpage/SelectPlaylistModal";
+// import LikedVideosPage from "./likedvideospage/LikedVideosPage";
+// import WatchLaterPage from "./watchlaterpage/WatchLaterPage";
+// import HistoryPage from "./historypage/HistoryPage";
+// import AddPlaylistModal from "./playlistpage/AddPlaylistModal";
+// import SelectPlaylistModal from "./playlistpage/SelectPlaylistModal";
+import VideoCard from "./videocard/VideoCard";
 
-export {NavBar,HomePage,Sidebar, LoginPage, SignUpPage, ExplorePage, RestrictedRoute, PrivateRoute, VideoPage, LikedVideosPage, WatchLaterPage, HistoryPage, AddPlaylistModal, SelectPlaylistModal}
+// export {NavBar,HomePage,Sidebar, LoginPage, SignUpPage, ExplorePage, RestrictedRoute, PrivateRoute, VideoPage, LikedVideosPage, WatchLaterPage, HistoryPage, AddPlaylistModal, SelectPlaylistModal}
+
+export {NavBar,Sidebar,RestrictedRoute,PrivateRoute,VideoCard}
