+++ src/pages/index.js
@@ -1,0 +1,10 @@
+
+
+export {default as LoginPage} from "./authpage/LoginPage"
+export {default as SignUpPage} from "./authpage/SignUpPage"
+export {default as ExplorePage} from "./explorepage/ExplorePage"
+export {default as HistoryPage} from "./historypage/HistoryPage"
+export {default as HomePage} from "./homepage/HomePage"
+export {default as LikedVideosPage} from "./likedVideospage/LikedVideosPage"
+export {default as PlaylistPage} from "./playlistpage/PlaylistPage"
+export {default as WatchLaterPage} from "./watchLaterpage/WatchLaterPage"
