+++ src/pages/videopage/VideoPage.jsx
@@ -1,10 +1,10 @@
 import React, { useEffect } from "react";
 import { useParams } from "react-router-dom";
+import { VideoCard } from "../../components";
 import { useAuth, useHistory, useLikes, usePlaylist, useVideos, useWatchLater } from "../../context";
 import { getSubscribersString, getViewString } from "../../utils";
 import AddPlaylistModal from "../playlistpage/AddPlaylistModal";
 import SelectPlaylistModal from "../playlistpage/SelectPlaylistModal";
-import VideoCard from "../videocard/VideoCard";
 import "./videopage.css";
 
 export default function VideoPage() {
